PostgreSQL - 2 POWER function , SQRT
You are playing a game on a rectangular checkerboard. Each game piece has a unique id and each cell on the board is defined by its x and y coordinates.

example

The current positions of your pieces on the board are stored in the positions table with the following structure:

id(SMALLINT): unique piece id;
x(SMALLINT): x coordinate of the cell that the piece with id id occupies;
y(SMALLINT): y coordinate of the cell that the piece with id id occupies.
In this game each piece on the board is said to defend its nearest neighbor. The distance between two pieces is calculated simply as the distance between the points (x1, y1) and (x2, y2), where (x1, y1) and (x2, y2) are the coordinates of the cells occupied by the first and the second piece, respectively. You thought it might be good idea to find what piece each of the game pieces defends.

Given the positions table, compose the resulting table with two columns: id1 and id2, such that on each row the piece with id id1 defends the piece with id2 (i.e. id2 is the closest to id1 piece).
The table should be sorted by the values of id1 in ascending order.

It's guaranteed that for each piece there is only one other piece closest to it.

Example
For the following table positions

id	x	y
1	0	0
2	2	0
3	4	3
4	2	1
the output should be

id1	id2
1	2
2	4
3	4
4	2
Execution time limit: 2 seconds (PostgreSQL)

Suggestion
Choose the appropriate math function below to solve this problem.

Theory
Function	Description	Example	Result
ABS	Calculate the absolute value of a number	ABS(-10)	10
CBRT	Calculate the cube root of a number	CBRT(8)	2
CEIL	Round a number up to the nearest integer, which is greater than or equal to number	CEIL(-12.8)	-12
CEILING	Same as CEIL		
DEGREES	Convert radians to degrees	DEGREES(0.8)	45.83662361
DIV	Return the integer quotient of two numeric values	DIV(8,3)	2
EXP	Return the exponential value in scientific notation of a number	EXP(1)	2.718281828
FLOOR	Round a number down to the nearest integer, which is less than or equal to number	FLOOR(10.6)	10
LN	Return the natural logarithm of a numeric value	LN(3)	1.098612289
LOG	Return the base 10 logarithm of a numeric value	LOG(1000)	3
LOG	Return the logarithm of a numeric value to a specified base	LOG(8, 64)	2
MOD	Divide the first parameter by the second one and return the remainder	MOD(10,4)	2
PI	Return the value of PI	PI()	3.141592654
POWER	Raise a numeric value to the power of a second numeric value	POWER(5, 3)	125
RADIANS	Convert degrees to radians	RADIANS(60)	1.047197551
ROUND	Round a number to the nearest integer or to a specified decimal places	ROUND(10.3)	10
SCALE	Return the number of decimal digits in the fractional part	SCALE(1.234)	3
SIGN	Return the sign (positive, negative) of a numeric value	SIGN(-1)	-1
SQRT	Return the square root of a numeric value	SQRT(3.0)	1.732050808
TRUNC	Truncate a numeric value to a whole number of  to the specified decimal places	TRUNC(12.3)	12
WIDTH_BUCKET	Assign values to buckets in an equi-width histogram.