PostgreSQL - 86
You work for an online store in which each item put up for sale gets a unique randomly generated id.

These item ids are stored in the items table with only one column:

id(SMALLINT): unique id of an item.
However, this system proved to be not very convenient to use in a number of queries that required consecutive ids. To solve this problem, you decided to generate new ids for the items using the following algorithm: the item with the smallest id would get 1 as a new id, the second smallest would get 2, and so on.

Given the items table, compose the resulting table with two columns: newId and oldId . The first column should contain the old item id, and the second one should contain the new id generated as described above. The table should be sorted by the newId in ascending order.

Example
For the following table items

id
1
12
23
42
the output should be

newId	oldId
1	1
2	12
3	23
4	42
Execution time limit: 2 seconds (PostgreSQL)

Theory
The ROW_NUMBER function is a window function that assigns a sequential integer to each row in a result set. The following illustrates the syntax of the ROW_NUMBER function:

ROW_NUMBER() OVER(     
    [PARTITION BY column_1, column_2,…]
    [ORDER BY column_3,column_4,…]
)
The set of rows on which the ROW_NUMBER function operates is called a window.

The PARTITION BY clause divides the window into smaller sets or partitions. If you specify the PARTITION BY clause, the row number for each partition starts with one and increments by one. Because the PARTITION BY clause is optional to the ROW_NUMBER function, therefore you can omit it, and ROW_NUMBER function will treat the whole window as a partition.

The ORDER BY clause inside the OVER clause determines the order in which the numbers are assigned.

SELECT
   ROW_NUMBER() OVER() AS row_number,
   product_id, product_name,  
FROM products;