PostgreSQL - 89 ROW_NUMBER and LAG
Task
A reigning champion is a champion who's won both the previous and current years' competitions. The summer_medals table records the prize information of the athletes in the Olympics championships of the competitions. Look for Weightlifting champions of the two adjacent Olympics to find out who the reigning champions are.

The return table has the column year - champion year, champion column - the champion's nationality is taken from the country column in the summer_medals table, column last_champion - the nationality of the champion of the previous Olympics. In case of the last_champion column has NULL value, fill it with 'N/A'. The returning table needs to be sorted ascending year by year.

With summer_medals table, we have structure like below:

year (SMALLINT) : the year that Olympic events occurred.
city (VARCHAR[50]) : the city that Olympic events occurred.
sport (VARCHAR[50]) : the sport in Olympics.
athlete (VARCHAR[50]) : name of the athlete.
country (VARCHAR[50]) : nationality of that athlete.
gender (VARCHAR[50]) : gender of athlete.
medal (VARCHAR[50]) : the medal that athlete achieved.
Example
For the data of the following table:

summer_medals

year	city	sport	athlete	country	gender	medal
1996	Atlanta	Weightlifting	MUTLU Halil	TUR	Men	Gold
1996	Atlanta	Aquatics	XIAO Hailiang	CHN	Men	Bronze
1996	Atlanta	Aquatics	SAUTIN Dmitry	RUS	Men	Gold
2000	Sydney	Weightlifting	LAVRENOV Sergei	BLR	Men	Bronze
2000	Sydney	Weightlifting	BOEVSKI Galabin	BUL	Men	Gold
2000	Sydney	Weightlifting	MARKOV Georgi	BUL	Men	Silver
2004	Athens	Weightlifting	PECHALOV Nikolay	CRO	Men	Bronze
2004	Athens	Weightlifting	ZHANG Guozheng	CHN	Men	Gold
The output should be:

year	champion	last_champion
1996	TUR	N/A
2000	BUL	TUR
2004	CHN	BUL
Execution time limit: 0.5 second (PosgreSQL)

Suggestion
The champion is the one who wins the gold medal.